fix: mv get_transaction method from analytics to wallet
