+++ go/sdk/docs/WalletApi.md
@@ -8,6 +8,7 @@ Method | HTTP request | Description
 [**WalletDeleteRulePost**](WalletApi.md#WalletDeleteRulePost) | **Post** /wallet/delete_rule | Create &amp; broadcast delete-config-rule [async call]
 [**WalletGetNewAddressPost**](WalletApi.md#WalletGetNewAddressPost) | **Post** /wallet/get_new_address | Get a new address from a given diversifier or generate randomly
 [**WalletGetPublicKeyPost**](WalletApi.md#WalletGetPublicKeyPost) | **Post** /wallet/get_public_key | Get wallet public key
+[**WalletGetTransactionsPost**](WalletApi.md#WalletGetTransactionsPost) | **Post** /wallet/get_transactions | Get details on past transactions
 [**WalletGetWalletBalancesPost**](WalletApi.md#WalletGetWalletBalancesPost) | **Post** /wallet/get_wallet_balances | Get wallets information
 [**WalletIssueAssetPost**](WalletApi.md#WalletIssueAssetPost) | **Post** /wallet/issue_asset | Issue assets [async call]
 [**WalletTransferAssetPost**](WalletApi.md#WalletTransferAssetPost) | **Post** /wallet/transfer_asset | Transfer assets [async call]
@@ -105,6 +106,32 @@ Name | Type | Description  | Notes
 ### Return type
 
 [**GetPublicKeyResponse**](GetPublicKeyResponse.md)
+
+### Authorization
+
+[ApiKeyAuth](../README.md#ApiKeyAuth)
+
+### HTTP request headers
+
+ - **Content-Type**: application/json
+ - **Accept**: application/json
+
+[[Back to top]](#) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to Model list]](../README.md#documentation-for-models) [[Back to README]](../README.md)
+
+# **WalletGetTransactionsPost**
+> GetTransactionsResponse WalletGetTransactionsPost(ctx, getTransactionsRequest)
+Get details on past transactions
+
+### Required Parameters
+
+Name | Type | Description  | Notes
+------------- | ------------- | ------------- | -------------
+ **ctx** | **context.Context** | context for authentication, logging, cancellation, deadlines, tracing, etc.
+  **getTransactionsRequest** | [**GetTransactionsRequest**](GetTransactionsRequest.md)|  | 
+
+### Return type
+
+[**GetTransactionsResponse**](GetTransactionsResponse.md)
 
 ### Authorization
