+++ js/sdk/docs/WalletApi.md
@@ -8,6 +8,7 @@ Method | HTTP request | Description
 [**walletDeleteRulePost**](WalletApi.md#walletDeleteRulePost) | **POST** /wallet/delete_rule | Create &amp; broadcast delete-config-rule [async call]
 [**walletGetNewAddressPost**](WalletApi.md#walletGetNewAddressPost) | **POST** /wallet/get_new_address | Get a new address from a given diversifier or generate randomly
 [**walletGetPublicKeyPost**](WalletApi.md#walletGetPublicKeyPost) | **POST** /wallet/get_public_key | Get wallet public key
+[**walletGetTransactionsPost**](WalletApi.md#walletGetTransactionsPost) | **POST** /wallet/get_transactions | Get details on past transactions
 [**walletGetWalletBalancesPost**](WalletApi.md#walletGetWalletBalancesPost) | **POST** /wallet/get_wallet_balances | Get wallets information
 [**walletIssueAssetPost**](WalletApi.md#walletIssueAssetPost) | **POST** /wallet/issue_asset | Issue assets [async call]
 [**walletTransferAssetPost**](WalletApi.md#walletTransferAssetPost) | **POST** /wallet/transfer_asset | Transfer assets [async call]
@@ -183,6 +184,51 @@ Name | Type | Description  | Notes
 ### Return type
 
 [**GetPublicKeyResponse**](GetPublicKeyResponse.md)
+
+### Authorization
+
+[ApiKeyAuth](../README.md#ApiKeyAuth)
+
+### HTTP request headers
+
+ - **Content-Type**: application/json
+ - **Accept**: application/json
+
+<a name="walletGetTransactionsPost"></a>
+# **walletGetTransactionsPost**
+> GetTransactionsResponse walletGetTransactionsPost(getTransactionsRequest)
+
+Get details on past transactions
+
+### Example
+```javascript
+var QedItAssetTransfers = require('qed-it-asset-transfers');
+var defaultClient = QedItAssetTransfers.ApiClient.instance;
+// Configure API key authorization: ApiKeyAuth
+var ApiKeyAuth = defaultClient.authentications['ApiKeyAuth'];
+ApiKeyAuth.apiKey = 'YOUR API KEY';
+// Uncomment the following line to set a prefix for the API key, e.g. "Token" (defaults to null)
+//ApiKeyAuth.apiKeyPrefix = 'Token';
+
+var apiInstance = new QedItAssetTransfers.WalletApi();
+var getTransactionsRequest = new QedItAssetTransfers.GetTransactionsRequest(); // GetTransactionsRequest | 
+apiInstance.walletGetTransactionsPost(getTransactionsRequest).then(function(data) {
+  console.log('API called successfully. Returned data: ' + data);
+}, function(error) {
+  console.error(error);
+});
+
+```
+
+### Parameters
+
+Name | Type | Description  | Notes
+------------- | ------------- | ------------- | -------------
+ **getTransactionsRequest** | [**GetTransactionsRequest**](GetTransactionsRequest.md)|  | 
+
+### Return type
+
+[**GetTransactionsResponse**](GetTransactionsResponse.md)
 
 ### Authorization
